# openapi.yaml
openapi: 3.0.3
info:
  title: SGS - Système de Gestion Scolaire API
  description: |
    API RESTful pour le Système de Gestion Scolaire (SGS).
    Cette spécification définit tous les endpoints nécessaires pour le backend.
    
    ## Authentification
    L'API utilise JWT Bearer tokens pour l'authentification.
    Tous les endpoints (sauf /auth/login) requièrent un header Authorization valide.
    
    ## Pagination
    Les endpoints de liste supportent la pagination avec les paramètres:
    - `page`: numéro de page (défaut: 1)
    - `per_page`: éléments par page (défaut: 20, max: 100)
    - `sort_by`: champ de tri
    - `sort_order`: asc ou desc
    - `search`: recherche textuelle
    
    ## Codes d'erreur
    - 400: Requête invalide
    - 401: Non authentifié
    - 403: Accès refusé
    - 404: Ressource non trouvée
    - 422: Erreur de validation
    - 500: Erreur serveur
  version: 1.0.0
  contact:
    name: Équipe SGS
    email: dev@sgs.edu
servers:
  - url: http://localhost:3001/api
    description: Serveur de développement
  - url: https://api.sgs.edu
    description: Serveur de production
...
paths:
...
  /eleves:
    get:
      tags: [Élèves]
      summary: Liste des élèves avec pagination
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/per_page'
        - $ref: '#/components/parameters/search'
        - name: classe_id
          in: query
          schema:
            type: string
        - name: niveau_id
          in: query
          schema:
            type: string
        - name: statut
          in: query
          schema:
            type: string
            enum: [inscrit, reinscrit, transfere, abandonne, diplome]
        - name: sexe
          in: query
          schema:
            type: string
            enum: [M, F]
      responses:
        '200':
          description: Liste paginée des élèves
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedEleves'
    post:
      tags: [Élèves]
      summary: Créer un nouvel élève
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EleveCreate'
      responses:
        '201':
          description: Élève créé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Eleve'
        '422':
          $ref: '#/components/responses/ValidationError'
  /eleves/{id}:
    get:
      tags: [Élèves]
      summary: Récupérer un élève par ID
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: Détails de l'élève
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Eleve'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Élèves]
      summary: Mettre à jour un élève
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EleveUpdate'
      responses:
        '200':
          description: Élève mis à jour
    delete:
      tags: [Élèves]
      summary: Supprimer un élève
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '204':
          description: Élève supprimé
...
  /comptabilite/factures/eleve/{id}:
    get:
      tags: [Comptabilité]
      summary: Situation financière d'un élève
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: Facture de l'élève
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FactureEleve'
...
  /rh/enseignants:
    get:
      tags: [RH]
      summary: Liste des enseignants
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/per_page'
        - $ref: '#/components/parameters/search'
        - name: statut
          in: query
          schema:
            type: string
            enum: [actif, conge, suspendu, demission, licencie]
      responses:
        '200':
          description: Liste paginée des enseignants
    post:
      tags: [RH]
      summary: Créer un enseignant
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnseignantCreate'
      responses:
        '201':
          description: Enseignant créé
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    id:
      name: id
      in: path
      required: true
      schema:
        type: string
    page:
      name: page
      in: query
      schema:
        type: integer
        default: 1
    per_page:
      name: per_page
      in: query
      schema:
        type: integer
        default: 20
        maximum: 100
    search:
      name: search
      in: query
      schema:
        type: string
  responses:
...
    NotFound:
      description: Ressource non trouvée
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ValidationError:
      description: Erreur de validation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
  schemas:
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
    ValidationError:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
            details:
              type: object
              additionalProperties:
                type: array
                items:
                  type: string
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
        per_page:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer
...
    Eleve:
      type: object
      properties:
        id:
          type: string
        matricule:
          type: string
        nom:
          type: string
        prenom:
          type: string
        date_naissance:
          type: string
          format: date
        lieu_naissance:
          type: string
        sexe:
          type: string
          enum: [M, F]
        nationalite:
          type: string
        telephone:
          type: string
        email:
          type: string
        statut:
          type: string
          enum: [inscrit, reinscrit, transfere, abandonne, diplome]
        classe_actuelle:
          $ref: '#/components/schemas/Classe'
        tuteurs:
          type: array
          items:
            $ref: '#/components/schemas/Tuteur'
    EleveCreate:
      type: object
      required: [nom, prenom, date_naissance, lieu_naissance, sexe, nationalite, tuteurs]
      properties:
        nom:
          type: string
        prenom:
          type: string
        date_naissance:
          type: string
          format: date
        lieu_naissance:
          type: string
        sexe:
          type: string
          enum: [M, F]
        nationalite:
          type: string
        telephone:
          type: string
        email:
          type: string
        adresse:
          type: string
        tuteurs:
          type: array
          items:
            $ref: '#/components/schemas/Tuteur'
    EleveUpdate:
      type: object
      properties:
        nom:
          type: string
        prenom:
          type: string
        telephone:
          type: string
        email:
          type: string
        adresse:
          type: string
    Tuteur:
      type: object
      required: [nom, prenom, telephone, lien_parente]
      properties:
        nom:
          type: string
        prenom:
          type: string
        telephone:
          type: string
        email:
          type: string
        profession:
          type: string
        lien_parente:
          type: string
          enum: [pere, mere, tuteur, autre]
    Classe:
      type: object
      properties:
        id:
          type: string
        nom:
          type: string
        niveau_id:
          type: string
        capacite:
          type: integer
        effectif:
          type: integer
...
    PaginatedEleves:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Eleve'
            meta:
              $ref: '#/components/schemas/PaginationMeta'
...
    FactureEleve:
      type: object
      properties:
        eleve_id:
          type: string
        total_du:
          type: number
        total_paye:
          type: number
        solde:
          type: number
        statut:
          type: string
          enum: [en_attente, partiel, complete, en_retard]
...
    EnseignantCreate:
      type: object
      required: [nom, prenom, email, telephone, sexe, specialite, date_embauche, type_contrat, salaire_base]
      properties:
        nom:
          type: string
        prenom:
          type: string
        email:
          type: string
          format: email
        telephone:
          type: string
        sexe:
          type: string
          enum: [M, F]
        specialite:
          type: string
        diplome:
          type: string
        date_embauche:
          type: string
          format: date
        type_contrat:
          type: string
          enum: [cdi, cdd, vacation, stage]
        salaire_base:
          type: number
        matiere_ids:
          type: array
          items:
            type: string
...